+++ client/wfnews-war/src/main/angular/src/app/components/public-incident-page/incident-info-header/incident-header-panel.component.ts
@@ -4,6 +4,8 @@ import {
   Input,
   AfterViewInit,
   HostListener,
+  Output,
+  EventEmitter,
 } from '@angular/core';
 import { EvacOrderOption } from '../../../conversion/models';
 import * as L from 'leaflet';
@@ -23,6 +25,8 @@ import { ActivatedRoute, ParamMap, Router } from '@angular/router';
 import { Location } from '@angular/common';
 import { LocationData } from '@app/components/wildfires-list-header/filter-by-location/filter-by-location-dialog.component';
 import { PublishedIncidentService } from '@app/services/published-incident-service';
+import { toCanvas } from 'qrcode';
+
 
 @Component({
   selector: 'incident-header-panel',
@@ -34,6 +38,7 @@ export class IncidentHeaderPanel implements AfterViewInit {
   @Input() public incident: any;
   @Input() public evacOrders: EvacOrderOption[] = [];
   @Input() public extent: any;
+  @Output() requestPrint = new EventEmitter<void>();
 
   public params: ParamMap;
   public defaultEvacURL: string;
@@ -89,13 +94,19 @@ export class IncidentHeaderPanel implements AfterViewInit {
     ];
     this.map = L.map('map', {
       attributionControl: false,
-      zoomControl: true,
+      zoomControl: false,
       dragging: false,
       doubleClickZoom: false,
       boxZoom: false,
       trackResize: false,
       scrollWheelZoom: false,
     }).setView(location, 9);
+    if (!this.isMobileView()){
+      // only apply these in desktop
+      L.control.zoom({
+        position: 'topright'
+      }).addTo(this.map);
+    }
     // configure map data
     L.tileLayer('https://tile.openstreetmap.org/{z}/{x}/{y}.png', {
       attribution:
@@ -225,9 +236,10 @@ export class IncidentHeaderPanel implements AfterViewInit {
     return formattedDate;
   }
 
-  openContactUsWindow() {
+  openContactUsWindow(mode:string | null) {
     this.dialog.open(ContactUsDialogComponent, {
       panelClass: 'contact-us-dialog',
+      width: mode === 'desktop' ? '500px' : undefined,  // Set width based on mode
       data: {
         fireCentre: convertToFireCentreDescription(
           this.incident.contactOrgUnitIdentifer ||
@@ -340,4 +352,30 @@ this.router.navigate([ResourcesRoutes.DASHBOARD]);
       );
     }
   }
+
+  // printPage(){
+  //   this.requestPrint.emit();
+  // }
+
+  public printPage() {
+    const printContents =
+      document.getElementsByClassName('page-container')[0].innerHTML;
+
+    const appRoot = document.body.removeChild(
+      document.getElementById('app-root'),
+    );
+
+    document.body.innerHTML = printContents;
+
+    const canvas = document.getElementById('qr-code');
+    toCanvas(canvas, window.location.href, function(error) {
+      if (error) {
+console.error(error);
+}
+      window.print();
+      document.body.innerHTML = '';
+      document.body.appendChild(appRoot);
+    });
+  }
+
 }
