+++ client/wfnews-war/src/main/angular/src/app/components/public-incident-page/incident-info-panel/incident-info-panel.component.html
@@ -8,9 +8,6 @@
         <img alt="Twitter Icon" src="/assets/images/svg-icons/twitter-x.svg" class="icon twitter">
       </a>
       <mat-icon (click)="copyToClipboard()" class="button no-print">share</mat-icon>
-      <button class="print-button no-print" (click)="printPage()"><span>Print</span>
-        <mat-icon class="button">print</mat-icon>
-      </button>
     </div>
   </div>
   <div class="content-panel">
