+++ client/wfnews-war/src/main/angular/src/app/components/public-incident-page/public-incident-page.component.scss
@@ -103,7 +103,6 @@
 }
 
 .page-container {
-  max-width: 1185px;
   margin: auto;
   background-color: white;
 }
